split code editor
